// commonwcl/projects/commonwcl/src/lib/models/dati-impresa.ts
export interface DatiImpresa {
    id_persona_giuridica?: number;
    denominazione?: string;
    codice_fiscale?: string;
    desStato?: string;
    indirizzo_sitad?: string;
    stradario?: number;
    indirizzo_non_trovato?: string;
    sigla_prov?: string;
    istat_comune?: string;
    comune?: string;
    provincia?: string;
    civico?: string;
    cap?: string;
    email?: string;
    data_inizio_attivita?: Date;
    data_cessazione?: Date;
    sigla_rea?: string;
    numero_rea?: number;
    flg_amministratore?: number;
    data_ult_mod?: Date;
    utente_ult_mod?: string;
    flg_terzo_responsabile?: number;
    flg_distributore?: number;
    flg_cat?: number;
    flg_indirizzo_estero?: number;
    stato_estero?: string;
    indirizzo_estero?: string;
    cap_estero?: string;
    fk_stato_pg?: number;
    dt_agg_dichiarazione?: Date;
    flg_dm37_letterac?: number;
    flg_dm37_letterad?: number;
    flg_dm37_letterae?: number;
    flg_fgas?: number;
    flg_conduttore?: number;
    flg_sogg_incaricato?: number;
    data_fine_legame?: Date;
    flg_delega?: number;
    delega_sogg_incaricato?: string;
    pec?: string;
    telefono?: string;
    token?: string;
    dt_creazione_token?: Date;
    dt_scadenza_token?: Date;
    citta_estero?: string;
}